requirements.txt
fastapi[standard]
SQLAlchemy
sqlmodel
psycopg2-binary
bcrypt
pydantic-extra-types[phonenumbers]
alembic
PyJWT
resend
pydantic-settings
fastapi-limiter
